move json response tests to separate module
